Removed MacOS, Emscripten and Android support
~ They weren't updated nor tested in a long time, so it's better to drop the support and to focus on
Desktop
~ Support might come back in the distant future

+++ core/src/utils/file_utility.cpp
@@ -33,136 +33,30 @@
 #include "easy/profiler.h"
 #endif
 
-#if defined(__ANDROID__)
-#include <android/asset_manager.h>
-#include <android/asset_manager_jni.h>
-#include <jni.h>
-
-static AAssetManager* assetManager = nullptr;
-
-extern "C"
-{
-	JNIEXPORT void JNICALL Java_swiss_sae_gpr5300_MainActivity_load(
-		JNIEnv* env, [[maybe_unused]] jclass clazz, jobject mgr)
-	{
-		assetManager = AAssetManager_fromJava(env, mgr);
-		logDebug("Asset Manager from JNI loaded!");
-	}
-}
-namespace neko
-{
-ResourceJob::ResourceJob() : Job([this] { bufferFile_.Load(filePath_); }) {}
-void ResourceJob::SetFilePath(std::string_view path) { filePath_ = path; }
-
-bool FileExists(const std::string_view path)
-{
-	{
-		std::ostringstream oss;
-		oss << "Checking file exists: " << path;
-		logDebug(oss.str());
-	}
-	AAsset* file = AAssetManager_open(assetManager, path.data(), AASSET_MODE_STREAMING);
-	bool exist   = file != nullptr;
-	if (exist) AAsset_close(file);
-	return exist;
-}
-std::string GetFilenameExtension(const std::string_view path)
-{
-	const std::string extension = path.data();
-	return extension.substr(extension.find_last_of("."));
-}
-std::string GetCurrentPath() { return ""; }
-const std::string LoadFile(const std::string& path)
-{
-	AAsset* file = AAssetManager_open(assetManager, path.c_str(), AASSET_MODE_BUFFER);
-	if (file == nullptr) return "";
-	// Get the file length
-	const size_t fileLength = static_cast<const size_t>(AAsset_getLength64(file));
-
-	// Allocate memory to read your file
-	char* fileContent = new char[fileLength + 1];
-
-	// Read your file
-	AAsset_read(file, fileContent, fileLength);
-	fileContent[fileLength] = '\0';
-	AAsset_close(file);
-	std::string str(fileContent);
-	delete[] fileContent;
-	return str;
-}
-
-void BufferFile::Load(std::string_view path)
-{
-	AAsset* file = AAssetManager_open(assetManager, path.data(), AASSET_MODE_BUFFER);
-	if (file == nullptr) return;
-	// Get the file length
-	dataLength = static_cast<const size_t>(AAsset_getLength64(file));
-
-	// Allocate memory to read your file
-	dataBuffer             = new unsigned char[dataLength + 1];
-	dataBuffer[dataLength] = '\0';
-
-	// Read your file
-	AAsset_read(file, dataBuffer, dataLength);
-	AAsset_close(file);
-}
-
-void BufferFile::Destroy()
-{
-	delete[] dataBuffer;
-	dataBuffer = nullptr;
-	dataLength = 0;
-}
-
-BufferFile::~BufferFile() { Destroy(); }
-
-void ResourceJob::Reset()
-{
-	Job::Reset();
-	bufferFile_.Destroy();
-}
-}    // namespace neko
-#else
-#ifdef __APPLE__
-#include <boost/filesystem.hpp>
-namespace fs = boost::filesystem;
-#elif WIN32
+#ifdef WIN32
 #include <filesystem>
 namespace fs = std::filesystem;
 #elif defined(__linux__) || defined(EMSCRIPTEN)
 #include <filesystem>
 namespace fs = std::filesystem;
-#endif
 
 namespace neko
 {
 bool FileExists(const std::string_view filename)
 {
-#ifdef __APPLE__
-	const fs::path p = std::string(filename);
-#else
 	const fs::path p = filename;
-#endif
 	return fs::exists(p);
 }
 
 bool IsRegularFile(const std::string_view filename)
 {
-#ifdef __APPLE__
-	const fs::path p = std::string(filename);
-#else
 	const fs::path p = filename;
-#endif
 	return fs::is_regular_file(p);
 }
 
 bool IsDirectory(const std::string_view filename)
 {
-#ifdef __APPLE__
-	const fs::path p = std::string(filename);
-#else
 	const fs::path p = filename;
-#endif
 	return fs::is_directory(p);
 }
 
@@ -176,31 +70,13 @@ std::string GetCurrentPath() { return fs::current_path().string(); }
 
 bool CreateDirectory(const std::string_view dirname)
 {
-#ifdef __APPLE__
-	return fs::create_directory(std::string(dirname));
-#else
 	return fs::create_directory(dirname);
-#endif
 }
 
 bool RemoveDirectory(const std::string_view dirname, bool removeAll)
 {
-	if (removeAll)
-	{
-#ifdef __APPLE__
-		return fs::remove_all(std::string(dirname));
-#else
-		return fs::remove_all(dirname);
-#endif
-	}
-	else
-	{
-#ifdef __APPLE__
-		return fs::remove(std::string(dirname));
-#else
-		return fs::remove(dirname);
-#endif
-	}
+	if (removeAll) return fs::remove_all(dirname);
+	else return fs::remove(dirname);
 }
 
 void IterateDirectory(const std::string_view dirname,
@@ -209,11 +85,7 @@ void IterateDirectory(const std::string_view dirname,
 {
 	if (IsDirectory(dirname))
 	{
-#ifdef __APPLE__
-		for (auto& p : fs::directory_iterator(std::string(dirname)))
-#else
 		for (auto& p : fs::directory_iterator(dirname))
-#endif
 		{
 			if (IsRegularFile(p.path().generic_string())) { func(p.path().generic_string()); }
 			else if (recursive && IsDirectory(p.path().generic_string()))
@@ -251,11 +123,7 @@ std::string GetRelativePath(const std::string_view path, const std::string_view 
 
 std::string GetFileParentPath(const std::string_view path)
 {
-#ifdef __APPLE__
-	fs::path p = std::string(path);
-#else
 	fs::path p       = fs::path(path);
-#endif
 	return p.parent_path().string();
 }
 
@@ -279,14 +147,8 @@ std::string GetFilenameExtension(const std::string_view path)
 
 std::string LinkFolderAndFile(const std::string_view folderPath, const std::string_view filePath)
 {
-
-#ifdef __APPLE__
-	fs::path f       = std::string(folderPath);
-	fs::path p       = std::string(filePath);
-#else
 	const fs::path f = fs::path(folderPath);
 	const fs::path p = fs::path(filePath);
-#endif
 	const fs::path l = f / p;
 
 	return MakeGeneric(l.string());
